Fix lints and flow errors for v0.10 (#856)
* Fix trailing comma lints

I'm not sure why ESLint didn't catch these at some point on Github - but
in any case we are fixing it now.

I ran the `trailing-commas.js` js-codemod and it only touched one file
for some reason. Probably I missed something about the jscodeshift API.
In any case, I fixed the other lints that our internal linter was
catching.

* Use DraftEntity type instead of EntityMap type in helpers

While supporting both the old and new APIs in Draft.js for Entity
manipulation, we are calling to the old DraftEntity module under the
hood instead of using the new system that would use an Immutable.js
OrderedMap called the EntityMap.

The DraftEntity module now has private methods for entity manipulation
that don't warn, and the public API which we are deprecating has
warnings and then calls the private methods.

In short, there are places where the flow typing wasn't right because
it was expecting EntityMap, and not DraftEntity.

I don't know why flow didn't catch this when we merged
https://github.com/facebook/draft-js/pull/814

test plan: `flow src`

* Fix flow error caused by passing command to RichUtils.handleKeyCommand

The 'handleKeyCommand' takes either a 'DraftEditorCommand' type or any
string.  This allows users to have their own custom editor command
types. I'm not sure how common this is, but our current flow typing
allows it.

The flow typing of 'RichUtils.handleKeyCommand' is more strict - it only
allows 'DraftEditorCommand' types, and not custom strings. This causes
an issue in one of our use cases, where we pass the 'command' received
in 'handleKeyCommand' through to 'RichUtils.handleKeyCommand'.

It seems like the 'RichUtils.handleKeyCommand' should have an API
similar to 'DraftEditor.handleKeyCommand', and so we're loosening the
flow type on 'RichTuils.handleKeyCommand'.

Test Plan: `flow`

* Fix flow error with DraftEditor.setClipboard

This was causing flow errors in our internal code because we expect that
'null' or 'undefined' is ok as an argument to 'setClipboard'.

Test Plan: `flow src`

* Fix flow and lint errors in CharacterMetadata

Again - I'm not sure why the github CI didn't catch these issues
earlier. I wonder if the issue with the typing of Immutable.OrderedMap
relates to the version of Immutable.js we are using or something.